feat: add env support, add Docker image

+++ internal/sender/telegram.go
@@ -15,7 +15,7 @@ type Sender struct {
 
 func NewSender(AppID int32, AppHash string, BotToken string) (*Sender, error) {
 	client, err := telegram.NewClient(telegram.ClientConfig{
-		AppID:    int32(AppID),
+		AppID:    AppID,
 		AppHash:  AppHash,
 		LogLevel: telegram.LogError,
 	})
